+++ bambi/__init__.py
@@ -1,3 +1,5 @@
+import logging
+
 from .models import Model
 from .priors import Prior, Family
 from .backends import PyMC3BackEnd
@@ -5,3 +7,11 @@ from .version import __version__
 
 
 __all__ = ["Model", "Prior", "Family", "PyMC3BackEnd"]
+
+_log = logging.getLogger("bambi")
+
+if not logging.root.handlers:
+    _log.setLevel(logging.INFO)
+    if len(_log.handlers) == 0:
+        handler = logging.StreamHandler()
+        _log.addHandler(handler)